+++ src/app/cart/shared/index.ts
@@ -1,2 +1,2 @@
-export { CartItemModel } from './models/cart.model';
+export * from './models';
 export { getIndex } from './helpers/get-index';