+++ src/app/cart/shared/models/index.ts
@@ -1,0 +1,3 @@
+export { SortOptionsModel } from './sort-options.model';
+export { CartItemModel } from './cart.model';
+export { SortConfigModel } from './sort-config.model';