+++ src/app/cart/shared/models/sort-options.model.ts
@@ -1,0 +1,4 @@
+export interface SortOptionsModel {
+  value: string;
+  label: string;
+}