+++ src/app/core/core.module.ts
@@ -1,8 +1,7 @@
-import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
 @NgModule({
-  imports: [CommonModule],
+  imports: [],
   declarations: []
 })
 export class CoreModule {}