+++ src/app/products/product-list/products.service.ts
@@ -1,18 +1,20 @@
 import { Injectable } from '@angular/core';
+import { BehaviorSubject, Observable } from 'rxjs';
 import { Categories, Product, ProductModel } from '../shared';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ProductsService {
-  constructor() { }
+  private products: BehaviorSubject<ProductModel[]> = new BehaviorSubject<ProductModel[]>([
+    new Product('1', 'Milk', 'Fresh milk', 1000, Categories.PRODUCTS, true),
+    new Product('2', 'Water', 'Pure mineral water', 10, Categories.DRINKS, true),
+    new Product('3', 'Jeans', 'Denim in your style', 1000, Categories.CLOTHES, true),
+    new Product('4', 'Meat', 'Fresh meat', 200, Categories.PRODUCTS, false)
+  ]);
+  constructor() {}
 
-  getProducts(): ProductModel[] {
-    return [
-      new Product('1', 'Milk', 'Fresh milk', 100, Categories.PRODUCTS, false),
-      new Product('2', 'Water', 'Pure mineral water', 10, Categories.DRINKS, true),
-      new Product('3', 'Jeans', 'Denim in your style', 1000, Categories.CLOTHES, true),
-      new Product('4', 'Meat', 'Fresh meat', 200, Categories.PRODUCTS, false)
-    ];
+  getProducts(): Observable<ProductModel[]> {
+    return this.products.asObservable();
   }
 }