+++ src/app/shared/index.ts
@@ -1,1 +1,2 @@
 export { SharedModule } from './shared.module';
+export * from './pipes/index';