+++ changelog.md
@@ -9,3 +9,11 @@
 - Добавлены сервисы для списка продуктов и корзины, которые пока будут возвращать соответствующие данные
 - Добавлен функционал отображения пустой корзины
 - Оптимизация отслеживания массива продутов для корзины с помощью директивы *ngFor + trackBy
+
+## [Task 4]
+### Added
+
+- titlecase pipe использован в product.component.html
+- titlecase pipe использован в cart-item.component.html
+- currency pipe использован в cart-item.component.html
+- orderBy pipe использован в cart-list.component.ts